catch core data

+++ Sources/EmbraceUploadInternal/Cache/EmbraceUploadCache.swift
@@ -125,7 +125,7 @@ class EmbraceUploadCache {
                 if let uploadData = try context.fetch(request).first {
                     uploadData.data = data
                     uploadData.payloadTypes = payloadTypes
-                    try context.save()
+                    coreData.save()
                     return true
                 }
             } catch {
@@ -145,12 +145,10 @@ class EmbraceUploadCache {
                 attemptCount: 0,
                 date: Date()
             ) {
-
-                do {
-                    try context.save()
-                    return true
-                } catch {
-                    context.delete(record)
+                coreData.performOperation { _ in
+                    if !coreData.saveIfNeeded() {
+                        context.delete(record)
+                    }
                 }
             }
             return false
@@ -177,7 +175,7 @@ class EmbraceUploadCache {
                     context.delete(uploadData)
                 }
 
-                try context.save()
+                coreData.save()
             }
         } catch {
             logger.error("error checking count limit:\n\(error.localizedDescription)")
@@ -214,11 +212,7 @@ class EmbraceUploadCache {
 
             uploadData.attemptCount = attemptCount
 
-            do {
-                try self?.coreData.context.save()
-            } catch {
-                self?.logger.warning("Error upading attempt count:\n\(error.localizedDescription)")
-            }
+            self?.coreData.save()
         }
     }
